move send_email to config
