+++ fence/config.py
@@ -3,6 +3,7 @@ from functools import wraps
 
 import bcrypt
 import flask
+import requests
 from werkzeug.datastructures import ImmutableMultiDict
 from yaml import safe_load as yaml_load
 import urllib.parse
@@ -12,6 +13,7 @@ from gen3config import Config
 
 from cdislogging import get_logger
 
+from fence.errors import NotFound
 from fence.utils import log_backoff_retry, log_backoff_giveup, exception_do_not_retry, generate_client_credentials, \
     logger
 from fence.models import Client, User, query_for_user
@@ -262,3 +264,49 @@ def get_SQLAlchemyDriver(db_conn_url):
     # TODO move userdatamodel code to Fence and remove dependencies to it
     SQLAlchemyDriver.setup_db = lambda _: None
     return SQLAlchemyDriver(db_conn_url)
+
+
+def send_email(from_email, to_emails, subject, text, smtp_domain):
+    """
+    Send email to group of emails using mail gun api.
+
+    https://app.mailgun.com/
+
+    Args:
+        from_email(str): from email
+        to_emails(list): list of emails to receive the messages
+        text(str): the text message
+        smtp_domain(dict): smtp domain server
+
+            {
+                "smtp_hostname": "smtp.mailgun.org",
+                "default_login": "postmaster@mailgun.planx-pla.net",
+                "api_url": "https://api.mailgun.net/v3/mailgun.planx-pla.net",
+                "smtp_password": "password", # pragma: allowlist secret
+                "api_key": "api key" # pragma: allowlist secret
+            }
+
+    Returns:
+        Http response
+
+    Exceptions:
+        KeyError
+
+    """
+    if smtp_domain not in config["GUN_MAIL"] or not config["GUN_MAIL"].get(
+        smtp_domain
+    ).get("smtp_password"):
+        raise NotFound(
+            "SMTP Domain '{}' does not exist in configuration for GUN_MAIL or "
+            "smtp_password was not provided. "
+            "Cannot send email.".format(smtp_domain)
+        )
+
+    api_key = config["GUN_MAIL"][smtp_domain].get("api_key", "")
+    email_url = config["GUN_MAIL"][smtp_domain].get("api_url", "") + "/messages"
+
+    return requests.post(
+        email_url,
+        auth=("api", api_key),
+        data={"from": from_email, "to": to_emails, "subject": subject, "text": text},
+    )
