+++ fence/utils.py
@@ -6,7 +6,6 @@ import json
 from random import SystemRandom
 import re
 import string
-import requests
 from urllib.parse import urlencode
 from urllib.parse import parse_qs, urlsplit, urlunsplit
 import sys
@@ -14,7 +13,7 @@ import sys
 from cdislogging import get_logger
 import flask
 
-from fence.errors import NotFound, UserError
+from fence.errors import UserError
 from authlib.oauth2.rfc6749.util import scope_to_list
 from authlib.oauth2.rfc6749.errors import InvalidScopeError
 
@@ -158,52 +157,6 @@ def split_url_and_query_params(url):
     query_params = parse_qs(query_string)
     url = urlunsplit((scheme, netloc, path, None, fragment))
     return url, query_params
-
-
-def send_email(from_email, to_emails, subject, text, smtp_domain):
-    """
-    Send email to group of emails using mail gun api.
-
-    https://app.mailgun.com/
-
-    Args:
-        from_email(str): from email
-        to_emails(list): list of emails to receive the messages
-        text(str): the text message
-        smtp_domain(dict): smtp domain server
-
-            {
-                "smtp_hostname": "smtp.mailgun.org",
-                "default_login": "postmaster@mailgun.planx-pla.net",
-                "api_url": "https://api.mailgun.net/v3/mailgun.planx-pla.net",
-                "smtp_password": "password", # pragma: allowlist secret
-                "api_key": "api key" # pragma: allowlist secret
-            }
-
-    Returns:
-        Http response
-
-    Exceptions:
-        KeyError
-
-    """
-    if smtp_domain not in config["GUN_MAIL"] or not config["GUN_MAIL"].get(
-        smtp_domain
-    ).get("smtp_password"):
-        raise NotFound(
-            "SMTP Domain '{}' does not exist in configuration for GUN_MAIL or "
-            "smtp_password was not provided. "
-            "Cannot send email.".format(smtp_domain)
-        )
-
-    api_key = config["GUN_MAIL"][smtp_domain].get("api_key", "")
-    email_url = config["GUN_MAIL"][smtp_domain].get("api_url", "") + "/messages"
-
-    return requests.post(
-        email_url,
-        auth=("api", api_key),
-        data={"from": from_email, "to": to_emails, "subject": subject, "text": text},
-    )
 
 
 def get_valid_expiration_from_request(
